requirements.txt
mysql-connector-python>=8.0.0
python-dateutil>=2.8.2
pillow>=9.0.0
